requirements.txt
numpy
joblib
pandas
pytest
imblearn
scikit-learn

httpx
fastapi
pydantic

sqlalchemy
psycopg2-binary
python-dotenv
